@@ -1,33 +1,32 @@
-"use strict";
+/**
+ * Module : Sparrow ajax
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-28 19:06:36
+ */
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.ajax = undefined;
-
-var _env = require("./env");
+import { env } from './env';
 
 var XmlHttp = {
 	get: "get",
 	post: "post",
 	reqCount: 4,
-	createXhr: function createXhr() {
+	createXhr: function () {
 		var xmlhttp = null;
 		/*if (window.XMLHttpRequest) {
     xmlhttp = new XMLHttpRequest();
   } else {
     xmlhttp = new ActiveXObject("Microsoft.XMLHTTP");
   }*/
-		if (_env.env.isIE8) {
+		if (env.isIE8) {
 			xmlhttp = new ActiveXObject("Microsoft.XMLHTTP"); //IE低版本创建XMLHTTP
-		} else if (_env.env.isIE) {
+		} else if (env.isIE) {
 			xmlhttp = new ActiveXObject("Msxml2.XMLHTTP"); //IE高版本创建XMLHTTP
 		} else if (window.XMLHttpRequest) {
 			xmlhttp = new XMLHttpRequest();
 		}
 		return xmlhttp;
 	},
-	ajax: function ajax(_json) {
+	ajax: function (_json) {
 		var url = _json["url"];
 		var callback = _json["success"];
 		var async = _json["async"] == undefined ? true : _json["async"];
@@ -71,7 +70,7 @@ var XmlHttp = {
 			XmlHttp.execBack(xmlhttp, callback, error);
 		}
 	},
-	execBack: function execBack(xmlhttp, callback, error) {
+	execBack: function (xmlhttp, callback, error) {
 		//if (xmlhttp.readyState == 4
 		if (xmlhttp.status == 200 || xmlhttp.status == 304 || xmlhttp.readyState == 4) {
 			callback(xmlhttp.responseText, xmlhttp.status, xmlhttp);
@@ -88,7 +87,7 @@ var XmlHttp = {
 			}
 		}
 	},
-	serializeUrl: function serializeUrl(url) {
+	serializeUrl: function (url) {
 		var cache = "cache=" + Math.random();
 		if (url.indexOf("?") > 0) {
 			url += "&" + cache;
@@ -97,7 +96,7 @@ var XmlHttp = {
 		}
 		return url;
 	},
-	serializeParams: function serializeParams(params) {
+	serializeParams: function (params) {
 		var ud = undefined;
 		if (ud == params || params == null || params == "") {
 			return null;
@@ -111,11 +110,7 @@ var XmlHttp = {
 		}
 		return params;
 	}
-}; /**
-    * Module : Sparrow ajax
-    * Author : Kvkens(yueming@yonyou.com)
-    * Date	  : 2016-07-28 19:06:36
-    */
+};
 
 var ajax = XmlHttp.ajax;
-exports.ajax = ajax;
+export { ajax };
