@@ -1,24 +1,12 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.core = undefined;
-
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; /**
-                                                                                                                                                                                                                                                                               * Module : Sparrow core context
-                                                                                                                                                                                                                                                                               * Author : Kvkens(yueming@yonyou.com)
-                                                                                                                                                                                                                                                                               * Date	  : 2016-07-28 13:52:19
-                                                                                                                                                                                                                                                                               */
-
-
-var _extend = require('./extend');
-
-var _util = require('./util');
-
-var _cookies = require('./cookies');
-
-var _enumerables = require('./enumerables');
+/**
+ * Module : Sparrow core context
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-28 13:52:19
+ */
+import { extend } from './extend';
+import { createShellObject } from './util';
+import { getCookie, setCookie } from './cookies';
+import { U_LANGUAGES, U_THEME, U_LOCALE, U_USERCODE } from './enumerables';
 
 var environment = {};
 /**
@@ -56,7 +44,7 @@ var maskerMeta = {
  * @return {environment}
  */
 fn.getEnvironment = function () {
-	return (0, _util.createShellObject)(environment);
+	return createShellObject(environment);
 };
 
 /**
@@ -64,8 +52,8 @@ fn.getEnvironment = function () {
  * @return {clientAttributes}
  */
 fn.getClientAttributes = function () {
-	var exf = function exf() {};
-	return (0, _util.createShellObject)(clientAttributes);
+	var exf = function () {};
+	return createShellObject(clientAttributes);
 };
 
 fn.setContextPath = function (contextPath) {
@@ -87,8 +75,8 @@ fn.setClientAttribute = function (k, v) {
  * @return {clientAttributes}
  */
 fn.getSessionAttributes = function () {
-	var exf = function exf() {};
-	return (0, _util.createShellObject)(sessionAttributes);
+	var exf = function () {};
+	return createShellObject(sessionAttributes);
 };
 
 /**
@@ -98,7 +86,7 @@ fn.getSessionAttributes = function () {
  */
 fn.setSessionAttribute = function (k, v) {
 	sessionAttributes[k] = v;
-	(0, _cookies.setCookie)("ISES_" + k, v);
+	setCookie("ISES_" + k, v);
 };
 
 /**
@@ -150,7 +138,7 @@ fn.setMaskerMeta = function (type, meta) {
 		getMetaFunc = type;
 	} else {
 		if (!maskerMeta[type]) maskerMeta[type] = meta;else {
-			if ((typeof meta === 'undefined' ? 'undefined' : _typeof(meta)) != 'object') maskerMeta[type] = meta;else for (var key in meta) {
+			if (typeof meta != 'object') maskerMeta[type] = meta;else for (var key in meta) {
 				maskerMeta[type][key] = meta[key];
 			}
 		}
@@ -160,24 +148,24 @@ fn.getMaskerMeta = function (type) {
 	if (typeof getMetaFunc == 'function') {
 		var meta = getMetaFunc.call(this);
 		return meta[type];
-	} else return (0, _extend.extend)({}, maskerMeta[type]);
+	} else return extend({}, maskerMeta[type]);
 };
-environment.languages = (0, _cookies.getCookie)(_enumerables.U_LANGUAGES) ? (0, _cookies.getCookie)(_enumerables.U_LANGUAGES).split(',') : navigator.language ? navigator.language : 'zh-CN';
+environment.languages = getCookie(U_LANGUAGES) ? getCookie(U_LANGUAGES).split(',') : navigator.language ? navigator.language : 'zh-CN';
 if (environment.languages == 'zh-cn') environment.languages = 'zh-CN';
 if (environment.languages == 'en-us') environment.languages = 'en-US';
 
-environment.theme = (0, _cookies.getCookie)(_enumerables.U_THEME);
-environment.locale = (0, _cookies.getCookie)(_enumerables.U_LOCALE);
+environment.theme = getCookie(U_THEME);
+environment.locale = getCookie(U_LOCALE);
 //environment.timezoneOffset = (new Date()).getTimezoneOffset()
-environment.usercode = (0, _cookies.getCookie)(_enumerables.U_USERCODE);
+environment.usercode = getCookie(U_USERCODE);
 //init session attribute
 document.cookie.replace(/ISES_(\w*)=([^;]*);?/ig, function (a, b, c) {
 	sessionAttributes[b] = c;
 });
 
-var Core = function Core() {};
+var Core = function () {};
 Core.prototype = fn;
 
 var core = new Core();
 
-exports.core = core;
+export { core };
