@@ -1,18 +1,15 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.getElementTop = exports.getElementLeft = exports.showPanelByEle = exports.getScroll = exports.getOffset = exports.makeModal = exports.makeDOM = exports.getZIndex = exports.getStyle = exports.wrap = exports.css = exports.closest = exports.toggleClass = exports.hasClass = exports.removeClass = exports.addClass = undefined;
-
-var _event = require('./event');
-
+/**
+ * Module : Sparrow dom
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-08-16 13:59:17
+ */
+import { on, stopEvent } from './event';
 /**
  * 元素增加指定样式
  * @param value
  * @returns {*}
  */
-var addClass = function addClass(element, value) {
+var addClass = function (element, value) {
 	if (element) {
 		if (typeof element.classList === 'undefined') {
 			if (u._addClass) {
@@ -32,12 +29,7 @@ var addClass = function addClass(element, value) {
  * @param {Object} element
  * @param {Object} value
  */
-/**
- * Module : Sparrow dom
- * Author : Kvkens(yueming@yonyou.com)
- * Date	  : 2016-08-16 13:59:17
- */
-var removeClass = function removeClass(element, value) {
+var removeClass = function (element, value) {
 	if (element) {
 		if (typeof element.classList === 'undefined') {
 			if (u._removeClass) {
@@ -56,7 +48,7 @@ var removeClass = function removeClass(element, value) {
  * @param {Object} element
  * @param {Object} value
  */
-var hasClass = function hasClass(element, value) {
+var hasClass = function (element, value) {
 	if (!element) return false;
 	if (element.nodeName && (element.nodeName === '#text' || element.nodeName === '#comment')) return false;
 	if (typeof element.classList === 'undefined') {
@@ -76,7 +68,7 @@ var hasClass = function hasClass(element, value) {
  * @param {Object} element
  * @param {Object} value
  */
-var toggleClass = function toggleClass(element, value) {
+var toggleClass = function (element, value) {
 	if (typeof element.classList === 'undefined') {
 		return u._toggleClass(element, value);
 	} else {
@@ -89,7 +81,7 @@ var toggleClass = function toggleClass(element, value) {
  * @param {Object} element
  * @param {Object} selector
  */
-var closest = function closest(element, selector) {
+var closest = function (element, selector) {
 	var tmp = element;
 	while (tmp != null && !hasClass(tmp, selector) && tmp != document.body) {
 		tmp = tmp.parentNode;
@@ -104,7 +96,7 @@ var closest = function closest(element, selector) {
  * @param {Object} csstext
  * @param {Object} val
  */
-var css = function css(element, csstext, val) {
+var css = function (element, csstext, val) {
 	//TO DO : 实现u.相关方法
 	if (csstext instanceof Object) {
 		for (var k in csstext) {
@@ -123,12 +115,12 @@ var css = function css(element, csstext, val) {
 	}
 };
 
-var wrap = function wrap(element, parent) {
+var wrap = function (element, parent) {
 	var p = makeDOM(parent);
 	element.parentNode.insertBefore(p, element);
 	p.appendChild(element);
 };
-var getStyle = function getStyle(element, key) {
+var getStyle = function (element, key) {
 	//不要在循环里用
 	var allCSS;
 	if (window.getComputedStyle) {
@@ -146,13 +138,13 @@ var globalZIndex;
 /**
  * 统一zindex值, 不同控件每次显示时都取最大的zindex，防止显示错乱
  */
-var getZIndex = function getZIndex() {
+var getZIndex = function () {
 	if (!globalZIndex) {
 		globalZIndex = 2000;
 	}
 	return globalZIndex++;
 };
-var makeDOM = function makeDOM(htmlString) {
+var makeDOM = function (htmlString) {
 	var tempDiv = document.createElement("div");
 	tempDiv.innerHTML = htmlString;
 	var _dom = tempDiv.children[0];
@@ -161,7 +153,7 @@ var makeDOM = function makeDOM(htmlString) {
 /**
  * element
  */
-var makeModal = function makeModal(element, parEle) {
+var makeModal = function (element, parEle) {
 	var overlayDiv = document.createElement('div');
 	$(overlayDiv).addClass('u-overlay');
 	overlayDiv.style.zIndex = getZIndex();
@@ -174,13 +166,13 @@ var makeModal = function makeModal(element, parEle) {
 	}
 
 	element.style.zIndex = getZIndex();
-	(0, _event.on)(overlayDiv, 'click', function (e) {
-		(0, _event.stopEvent)(e);
+	on(overlayDiv, 'click', function (e) {
+		stopEvent(e);
 	});
 	return overlayDiv;
 };
 
-var getOffset = function getOffset(Node, offset) {
+var getOffset = function (Node, offset) {
 	if (!offset) {
 		offset = {};
 		offset.top = 0;
@@ -193,7 +185,7 @@ var getOffset = function getOffset(Node, offset) {
 	offset.left += Node.offsetLeft;
 	if (Node.offsetParent) return getOffset(Node.offsetParent, offset);else return offset;
 };
-var getScroll = function getScroll(Node, offset) {
+var getScroll = function (Node, offset) {
 	if (!offset) {
 		offset = {};
 		offset.top = 0;
@@ -211,7 +203,7 @@ var getScroll = function getScroll(Node, offset) {
 
 	if (Node.parentNode) return getScroll(Node.parentNode, offset);else return offset;
 };
-var showPanelByEle = function showPanelByEle(obj) {
+var showPanelByEle = function (obj) {
 	var ele = obj.ele,
 	    panel = obj.panel,
 	    position = obj.position,
@@ -268,7 +260,7 @@ var showPanelByEle = function showPanelByEle(obj) {
 	panel.style.top = top + 'px';
 };
 
-var getElementLeft = function getElementLeft(element) {
+var getElementLeft = function (element) {
 	var actualLeft = element.offsetLeft;
 	var current = element.offsetParent;
 	while (current !== null) {
@@ -282,7 +274,7 @@ var getElementLeft = function getElementLeft(element) {
 	}
 	return actualLeft - elementScrollLeft;
 };
-var getElementTop = function getElementTop(element) {
+var getElementTop = function (element) {
 	var actualTop = element.offsetTop;
 	var current = element.offsetParent;
 	while (current !== null) {
@@ -296,19 +288,4 @@ var getElementTop = function getElementTop(element) {
 	}
 	return actualTop - elementScrollTop;
 };
-exports.addClass = addClass;
-exports.removeClass = removeClass;
-exports.hasClass = hasClass;
-exports.toggleClass = toggleClass;
-exports.closest = closest;
-exports.css = css;
-exports.wrap = wrap;
-exports.getStyle = getStyle;
-exports.getZIndex = getZIndex;
-exports.makeDOM = makeDOM;
-exports.makeModal = makeModal;
-exports.getOffset = getOffset;
-exports.getScroll = getScroll;
-exports.showPanelByEle = showPanelByEle;
-exports.getElementLeft = getElementLeft;
-exports.getElementTop = getElementTop;
+export { addClass, removeClass, hasClass, toggleClass, closest, css, wrap, getStyle, getZIndex, makeDOM, makeModal, getOffset, getScroll, showPanelByEle, getElementLeft, getElementTop };
