@@ -1,115 +1,102 @@
-'use strict';
+/**
+ * Module : Sparrow entry index
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-08-04 09:48:36
+ */
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.u = undefined;
+import { extend } from './extend';
+import { setCookie, getCookie } from './cookies';
+import { createShellObject, execIgnoreError, getFunction, getJSObject, isDate, isNumber, isArray, isEmptyObject, inArray, isDomElement, each } from './util';
+import { env } from './env';
+import { on, off, trigger, stopEvent, event } from './event';
+import { addClass, removeClass, hasClass, toggleClass, closest, css, wrap, getStyle, getZIndex, makeDOM, makeModal, getOffset, getScroll, showPanelByEle } from './dom';
+import { Class } from './class';
+import { core } from './core';
 
-var _extend = require('./extend');
+import { ajax } from './ajax';
 
-var _cookies = require('./cookies');
+import { floatRender, integerRender, dateRender, dateTimeRender, timeRender, percentRender, dateToUTCString } from './util/dataRender';
 
-var _util = require('./util');
+import { NumberFormater, DateFormater } from './util/formater';
 
-var _env = require('./env');
+import { date } from './util/dateUtils';
+import { AddressMasker, NumberMasker, CurrencyMasker, PercentMasker } from './util/masker';
 
-var _event = require('./event');
+import { hotkeys } from './util/hotKeys';
 
-var _dom = require('./dom');
+import { Ripple } from './util/ripple';
 
-var _class = require('./class');
+import { RSAUtils, BigInt, BarrettMu, twoDigit } from './util/rsautils';
 
-var _core = require('./core');
-
-var _ajax = require('./ajax');
-
-var _dataRender = require('./util/dataRender');
-
-var _formater = require('./util/formater');
-
-var _dateUtils = require('./util/dateUtils');
-
-var _masker = require('./util/masker');
-
-var _hotKeys = require('./util/hotKeys');
-
-var _ripple = require('./util/ripple');
-
-var _rsautils = require('./util/rsautils');
-
-var _i18n = require('./util/i18n');
+import { trans } from './util/i18n';
 
 //公开接口、属性对外暴露
-var api = {
-	ajax: _ajax.ajax,
-	extend: _extend.extend,
-	setCookie: _cookies.setCookie,
-	getCookie: _cookies.getCookie,
-	createShellObject: _util.createShellObject,
-	execIgnoreError: _util.execIgnoreError,
-	getFunction: _util.getFunction,
-	getJSObject: _util.getJSObject,
-	isDate: _util.isDate,
-	isNumber: _util.isNumber,
-	isArray: _util.isArray,
-	isEmptyObject: _util.isEmptyObject,
-	inArray: _util.inArray,
-	isDomElement: _util.isDomElement,
-	each: _util.each,
-	on: _event.on,
-	off: _event.off,
-	trigger: _event.trigger,
-	stopEvent: _event.stopEvent,
-	event: _event.event,
-	addClass: _dom.addClass,
-	removeClass: _dom.removeClass,
-	hasClass: _dom.hasClass,
-	toggleClass: _dom.toggleClass,
-	closest: _dom.closest,
-	css: _dom.css,
-	wrap: _dom.wrap,
-	getStyle: _dom.getStyle,
-	getZIndex: _dom.getZIndex,
-	makeDOM: _dom.makeDOM,
-	makeModal: _dom.makeModal,
-	getOffset: _dom.getOffset,
-	getScroll: _dom.getScroll,
-	showPanelByEle: _dom.showPanelByEle,
-	Class: _class.Class,
-	core: _core.core,
-	floatRender: _dataRender.floatRender,
-	integerRender: _dataRender.integerRender,
-	dateRender: _dataRender.dateRender,
-	dateTimeRender: _dataRender.dateTimeRender,
-	timeRender: _dataRender.timeRender,
-	percentRender: _dataRender.percentRender,
-	dateToUTCString: _dataRender.dateToUTCString,
-	date: _dateUtils.date,
-	NumberFormater: _formater.NumberFormater,
-	DateFormater: _formater.DateFormater,
-	AddressMasker: _masker.AddressMasker,
-	NumberMasker: _masker.NumberMasker,
-	CurrencyMasker: _masker.CurrencyMasker,
-	PercentMasker: _masker.PercentMasker,
-	hotkeys: _hotKeys.hotkeys,
-	Ripple: _ripple.Ripple,
-	RSAUtils: _rsautils.RSAUtils,
-	BigInt: _rsautils.BigInt,
-	BarrettMu: _rsautils.BarrettMu,
-	twoDigit: _rsautils.twoDigit,
-	trans: _i18n.trans
-}; /**
-    * Module : Sparrow entry index
-    * Author : Kvkens(yueming@yonyou.com)
-    * Date	  : 2016-08-04 09:48:36
-    */
-
-(0, _extend.extend)(api, _env.env);
+let api = {
+	ajax: ajax,
+	extend: extend,
+	setCookie: setCookie,
+	getCookie: getCookie,
+	createShellObject: createShellObject,
+	execIgnoreError: execIgnoreError,
+	getFunction: getFunction,
+	getJSObject: getJSObject,
+	isDate: isDate,
+	isNumber: isNumber,
+	isArray: isArray,
+	isEmptyObject: isEmptyObject,
+	inArray: inArray,
+	isDomElement: isDomElement,
+	each: each,
+	on: on,
+	off: off,
+	trigger: trigger,
+	stopEvent: stopEvent,
+	event: event,
+	addClass: addClass,
+	removeClass: removeClass,
+	hasClass: hasClass,
+	toggleClass: toggleClass,
+	closest: closest,
+	css: css,
+	wrap: wrap,
+	getStyle: getStyle,
+	getZIndex: getZIndex,
+	makeDOM: makeDOM,
+	makeModal: makeModal,
+	getOffset: getOffset,
+	getScroll: getScroll,
+	showPanelByEle: showPanelByEle,
+	Class: Class,
+	core: core,
+	floatRender: floatRender,
+	integerRender: integerRender,
+	dateRender: dateRender,
+	dateTimeRender: dateTimeRender,
+	timeRender: timeRender,
+	percentRender: percentRender,
+	dateToUTCString: dateToUTCString,
+	date: date,
+	NumberFormater: NumberFormater,
+	DateFormater: DateFormater,
+	AddressMasker: AddressMasker,
+	NumberMasker: NumberMasker,
+	CurrencyMasker: CurrencyMasker,
+	PercentMasker: PercentMasker,
+	hotkeys: hotkeys,
+	Ripple: Ripple,
+	RSAUtils: RSAUtils,
+	BigInt: BigInt,
+	BarrettMu: BarrettMu,
+	twoDigit: twoDigit,
+	trans: trans
+};
+extend(api, env);
 // export api;
 //export default api;
-(0, _extend.extend)(api, window.u || {});
+extend(api, window.u || {});
 
 window.u = api;
 window.iweb = {};
 window.iweb.browser = window.u;
-exports.u = api;
+window.sparrow = window.u;
+export { u, sparrow };
