@@ -1,22 +1,13 @@
-'use strict';
+/**
+ * Module : Sparrow date util
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-08-06 13:37:20
+ */
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.date = undefined;
-
-var _core = require('../core');
-
-var _util = require('../util');
-
-var _i18n = require('./i18n');
-
-var u = {}; /**
-             * Module : Sparrow date util
-             * Author : Kvkens(yueming@yonyou.com)
-             * Date	  : 2016-08-06 13:37:20
-             */
-
+import { core } from '../core';
+import { dateFormat } from '../util';
+import { trans } from './i18n';
+var u = {};
 u.date = {
 
 	/**
@@ -40,17 +31,17 @@ u.date = {
 		}
 	},
 	_jsonLocale: {
-		months: (0, _i18n.trans)('date.months', "一月\n二月\n三月\n四月\n五月\n六月\n七月\n八月\n九月\n十月\n十一月\n十二月").split('\n'),
-		monthsShort: (0, _i18n.trans)('date.monthsShort', "1月\n2月\n3月\n4月\n5月\n6月\n7月\n8月\n9月\n10月\n11月\n12月").split('\n'),
-		weekdays: (0, _i18n.trans)('date.weekdays', "星期日\n星期一\n星期二\n星期三\n星期四\n星期五\n星期六").split('\n'),
-		weekdaysShort: (0, _i18n.trans)('date.weekdaysShort', "周日\n周一\n周二\n周三\n周四\n周五\n周六").split('\n'),
-		weekdaysMin: (0, _i18n.trans)('date.weekdaysMin', "日\n一\n二\n三\n四\n五\n六").split('\n'),
+		months: trans('date.months', "一月\n二月\n三月\n四月\n五月\n六月\n七月\n八月\n九月\n十月\n十一月\n十二月").split('\n'),
+		monthsShort: trans('date.monthsShort', "1月\n2月\n3月\n4月\n5月\n6月\n7月\n8月\n9月\n10月\n11月\n12月").split('\n'),
+		weekdays: trans('date.weekdays', "星期日\n星期一\n星期二\n星期三\n星期四\n星期五\n星期六").split('\n'),
+		weekdaysShort: trans('date.weekdaysShort', "周日\n周一\n周二\n周三\n周四\n周五\n周六").split('\n'),
+		weekdaysMin: trans('date.weekdaysMin', "日\n一\n二\n三\n四\n五\n六").split('\n'),
 		defaultMonth: ["1月", "2月", "3月", "4月", "5月", "6月", "7月", "8月", "9月", "10月", "11月", "12月"]
 	},
 
 	_formattingTokens: /(\[[^\[]*\])|(\\)?(Mo|MM?M?M?|Do|DDDo|DD?D?D?|ddd?d?|do?|w[o|w]?|W[o|W]?|Q|YYYY|YY|e|E|a|A|hh?|HH?|mm?|ss?|S{1,4}|x|X|zz?|ZZ?|.)/g,
 
-	leftZeroFill: function leftZeroFill(number, targetLength, forceSign) {
+	leftZeroFill: function (number, targetLength, forceSign) {
 		var output = '' + Math.abs(number),
 		    sign = number >= 0;
 		while (output.length < targetLength) {
@@ -61,59 +52,59 @@ u.date = {
 
 	_formats: {
 		//year
-		YY: function YY(date) {
+		YY: function (date) {
 			return u.date.leftZeroFill(date.getFullYear() % 100, 2);
 		},
-		YYYY: function YYYY(date) {
+		YYYY: function (date) {
 			return date.getFullYear();
 		},
 		//month
-		M: function M(date) {
+		M: function (date) {
 			return date.getMonth() + 1;
 		},
-		MM: function MM(date) {
+		MM: function (date) {
 			var m = u.date._formats.M(date);
 			return u.date.leftZeroFill(m, 2);
 		},
-		MMM: function MMM(date, language) {
+		MMM: function (date, language) {
 			var m = date.getMonth();
 			// return u.date._dateLocale[language].monthsShort[m];
 			return u.date._jsonLocale.monthsShort[m];
 		},
-		MMMM: function MMMM(date, language) {
+		MMMM: function (date, language) {
 			var m = date.getMonth();
 			// return u.date._dateLocale[language].months[m];
 			return u.date._jsonLocale.months[m];
 		},
 		//date
-		D: function D(date) {
+		D: function (date) {
 			return date.getDate();
 		},
-		DD: function DD(date) {
+		DD: function (date) {
 			var d = u.date._formats.D(date);
 			return u.date.leftZeroFill(d, 2);
 		},
 		// weekday
-		d: function d(date) {
+		d: function (date) {
 			return date.getDay();
 		},
-		dd: function dd(date, language) {
+		dd: function (date, language) {
 			var d = u.date._formats.d(date);
 			// return u.date._dateLocale[language].weekdaysMin[d];
 			return u.date._jsonLocale.weekdaysMin[d];
 		},
-		ddd: function ddd(date, language) {
+		ddd: function (date, language) {
 			var d = u.date._formats.d(date);
 			// return u.date._dateLocale[language].weekdaysShort[d];
 			return u.date._jsonLocale.weekdaysShort[d];
 		},
-		dddd: function dddd(date, language) {
+		dddd: function (date, language) {
 			var d = u.date._formats.d(date);
 			// return u.date._dateLocale[language].weekdays[d];
 			return u.date._jsonLocale.weekdays[d];
 		},
 		// am pm
-		a: function a(date) {
+		a: function (date) {
 			if (date.getHours() > 12) {
 				return 'pm';
 			} else {
@@ -121,33 +112,33 @@ u.date = {
 			}
 		},
 		//hour
-		h: function h(date) {
+		h: function (date) {
 			var h = date.getHours();
 			h = h > 12 ? h - 12 : h;
 			return h;
 		},
-		hh: function hh(date) {
+		hh: function (date) {
 			var h = u.date._formats.h(date);
 			return u.date.leftZeroFill(h, 2);
 		},
-		H: function H(date) {
+		H: function (date) {
 			return date.getHours();
 		},
-		HH: function HH(date) {
+		HH: function (date) {
 			return u.date.leftZeroFill(date.getHours(), 2);
 		},
 		// minutes
-		m: function m(date) {
+		m: function (date) {
 			return date.getMinutes();
 		},
-		mm: function mm(date) {
+		mm: function (date) {
 			return u.date.leftZeroFill(date.getMinutes(), 2);
 		},
 		//seconds
-		s: function s(date) {
+		s: function (date) {
 			return date.getSeconds();
 		},
-		ss: function ss(date) {
+		ss: function (date) {
 			return u.date.leftZeroFill(date.getSeconds(), 2);
 		}
 	},
@@ -157,7 +148,7 @@ u.date = {
   * @param date
   * @param formatString
   */
-	format: function format(date, formatString, language) {
+	format: function (date, formatString, language) {
 		if (!date) return ''; // renturn date 改为 return '',因：setFormat初始会赋值为undefined,造成二次选择报错
 		var array = formatString.match(u.date._formattingTokens),
 		    i,
@@ -165,7 +156,7 @@ u.date = {
 		    output = '';
 		var _date = u.date.getDateObj(date);
 		if (!_date) return date;
-		language = language || _core.core.getLanguages();
+		language = language || core.getLanguages();
 		for (i = 0, length = array.length; i < length; i++) {
 			if (u.date._formats[array[i]]) {
 				output += u.date._formats[array[i]](_date, language);
@@ -176,7 +167,7 @@ u.date = {
 		return output;
 	},
 
-	_addOrSubtract: function _addOrSubtract(date, period, value, isAdding) {
+	_addOrSubtract: function (date, period, value, isAdding) {
 		var times = date.getTime(),
 		    d = date.getDate(),
 		    m = date.getMonth(),
@@ -209,16 +200,16 @@ u.date = {
 		return _date;
 	},
 
-	add: function add(date, period, value) {
+	add: function (date, period, value) {
 		return u.date._addOrSubtract(date, period, value, 1);
 	},
-	sub: function sub(date, period, value) {
+	sub: function (date, period, value) {
 		return u.date._addOrSubtract(date, period, value, -1);
 	},
-	getDateObj: function getDateObj(value) {
+	getDateObj: function (value) {
 		if (!value || typeof value == 'undefined') return value;
 		var dateFlag = false;
-		var _date = new Date((0, _util.dateFormat)(value));
+		var _date = new Date(dateFormat(value));
 		if (isNaN(_date)) {
 			// IE的话对"2016-2-13 12:13:22"进行处理
 			var index1, index2, index3, s1, s2, s3, s4;
@@ -272,4 +263,4 @@ u.date = {
 };
 
 var date = u.date;
-exports.date = date;
+export { date };
