@@ -1,22 +1,13 @@
-'use strict';
+/**
+ * Module : Sparrow hotKeys
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-28 20:25:39
+ */
 
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-exports.hotkeys = undefined;
-
-var _class = require('../class');
-
-var _extend = require('../extend');
-
-var _util = require('../util');
-
-var hotkeys = {}; /**
-                   * Module : Sparrow hotKeys
-                   * Author : Kvkens(yueming@yonyou.com)
-                   * Date	  : 2016-07-28 20:25:39
-                   */
-
+import { isFunction } from '../class';
+import { extend } from '../extend';
+import { getFunction } from '../util';
+var hotkeys = {};
 hotkeys.special_keys = {
     27: 'esc', 9: 'tab', 32: 'space', 13: 'enter', 8: 'backspace', 145: 'scroll', 20: 'capslock',
     144: 'numlock', 19: 'pause', 45: 'insert', 36: 'home', 46: 'del', 35: 'end', 33: 'pageup',
@@ -31,18 +22,18 @@ hotkeys.shift_nums = {
 };
 
 hotkeys.add = function (combi, options, callback) {
-    if ((0, _class.isFunction)(options)) {
+    if (isFunction(options)) {
         callback = options;
         options = {};
     }
     var opt = {},
         defaults = { type: 'keydown', propagate: false, disableInInput: false, target: document.body, checkParent: true },
         that = this;
-    opt = (0, _extend.extend)(opt, defaults, options || {});
+    opt = extend(opt, defaults, options || {});
     combi = combi.toLowerCase();
 
     // inspect if keystroke matches
-    var inspector = function inspector(event) {
+    var inspector = function (event) {
         //event = $.event.fix(event); // jQuery event normalization.
         var element = this; //event.target;
         // @ TextNode -> nodeType == 3
@@ -159,4 +150,4 @@ hotkeys.scan = function (element, target) {
 
 var hotkeys = hotkeys;
 
-exports.hotkeys = hotkeys;
+export { hotkeys };
