@@ -1,17 +1,10 @@
-"use strict";
+/**
+ * Module : Sparrow abstract formater class
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-28 19:35:26
+ */
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.PhoneNumberMasker = exports.PercentMasker = exports.CurrencyMasker = exports.NumberMasker = exports.AddressMasker = undefined;
-
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; /**
-                                                                                                                                                                                                                                                                               * Module : Sparrow abstract formater class
-                                                                                                                                                                                                                                                                               * Author : Kvkens(yueming@yonyou.com)
-                                                                                                                                                                                                                                                                               * Date	  : 2016-07-28 19:35:26
-                                                                                                                                                                                                                                                                               */
-
-var _extend = require("../extend");
+import { extend } from '../extend';
 
 function AbstractMasker() {};
 
@@ -211,7 +204,7 @@ function AddressMasker(formatMeta) {
 };
 
 AddressMasker.prototype.update = function (formatMeta) {
-	this.formatMeta = (0, _extend.extend)({}, AddressMasker.DefaultFormatMeta, formatMeta);
+	this.formatMeta = extend({}, AddressMasker.DefaultFormatMeta, formatMeta);
 };
 
 AddressMasker.prototype.getExpress = function () {
@@ -253,7 +246,7 @@ AddressMasker.prototype.getVarElement = function (express) {
 		return obj.postcode;
 	};
 
-	if (_typeof(ex.getValue) == undefined) return new StringElement(express);else return ex;
+	if (typeof ex.getValue == undefined) return new StringElement(express);else return ex;
 };
 
 AddressMasker.prototype.formatArgument = function (obj) {
@@ -283,7 +276,7 @@ function NumberMasker(formatMeta) {
 };
 
 NumberMasker.prototype.update = function (formatMeta) {
-	this.formatMeta = (0, _extend.extend)({}, NumberMasker.DefaultFormatMeta, formatMeta);
+	this.formatMeta = extend({}, NumberMasker.DefaultFormatMeta, formatMeta);
 };
 
 /**
@@ -372,7 +365,7 @@ function CurrencyMasker(formatMeta) {
 };
 
 CurrencyMasker.prototype.update = function (formatMeta) {
-	this.formatMeta = (0, _extend.extend)({}, CurrencyMasker.DefaultFormatMeta, formatMeta);
+	this.formatMeta = extend({}, CurrencyMasker.DefaultFormatMeta, formatMeta);
 };
 
 /**
@@ -396,7 +389,7 @@ function PercentMasker(formatMeta) {
 };
 
 PercentMasker.prototype.update = function (formatMeta) {
-	this.formatMeta = (0, _extend.extend)({}, NumberMasker.DefaultFormatMeta, formatMeta);
+	this.formatMeta = extend({}, NumberMasker.DefaultFormatMeta, formatMeta);
 };
 
 PercentMasker.prototype.formatArgument = function (obj) {
@@ -478,7 +471,7 @@ PhoneNumberMasker.prototype = new NumberMasker();
 PhoneNumberMasker.prototype.formatMeta = null;
 
 PhoneNumberMasker.prototype.update = function (formatMeta) {
-	this.formatMeta = (0, _extend.extend)({}, PhoneNumberMasker.DefaultFormatMeta, formatMeta);
+	this.formatMeta = extend({}, PhoneNumberMasker.DefaultFormatMeta, formatMeta);
 };
 
 PhoneNumberMasker.prototype.formatArgument = function (obj) {
@@ -503,7 +496,7 @@ NumberMasker.DefaultFormatMeta = {
 	negativeFormat: "-n"
 };
 
-CurrencyMasker.DefaultFormatMeta = (0, _extend.extend)({}, NumberMasker.DefaultFormatMeta, {
+CurrencyMasker.DefaultFormatMeta = extend({}, NumberMasker.DefaultFormatMeta, {
 	//curSymbol: "",
 	positiveFormat: "n",
 	negativeFormat: "-n"
@@ -516,8 +509,4 @@ AddressMasker.defaultFormatMeta = {
 
 PhoneNumberMasker.defaultFormatMeta = {};
 
-exports.AddressMasker = AddressMasker;
-exports.NumberMasker = NumberMasker;
-exports.CurrencyMasker = CurrencyMasker;
-exports.PercentMasker = PercentMasker;
-exports.PhoneNumberMasker = PhoneNumberMasker;
+export { AddressMasker, NumberMasker, CurrencyMasker, PercentMasker, PhoneNumberMasker };
