@@ -1,18 +1,15 @@
-'use strict';
+/**
+ * Module : Sparrow ripple
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-29 08:42:13
+ */
 
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.URipple = exports.Ripple = undefined;
-
-var _env = require('../env');
-
-var _dom = require('../dom');
-
-var _event = require('../event');
+import { env } from '../env';
+import { addClass, removeClass } from '../dom';
+import { on, off } from '../event';
 
 var URipple = function URipple(element) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   this._element = element;
 
   // Initialize instance.
@@ -20,14 +17,8 @@ var URipple = function URipple(element) {
 };
 //window['URipple'] = URipple;
 
-/**
- * Module : Sparrow ripple
- * Author : Kvkens(yueming@yonyou.com)
- * Date	  : 2016-07-29 08:42:13
- */
-
 URipple.prototype._down = function (event) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   if (!this._rippleElement.style.width && !this._rippleElement.style.height) {
     var rect = this._element.getBoundingClientRect();
     this.rippleSize_ = Math.sqrt(rect.width * rect.width + rect.height * rect.height) * 2 + 2;
@@ -37,7 +28,7 @@ URipple.prototype._down = function (event) {
     }
   }
 
-  (0, _dom.addClass)(this._rippleElement, 'is-visible');
+  addClass(this._rippleElement, 'is-visible');
 
   if (event.type === 'mousedown' && this._ignoringMouseDown) {
     this._ignoringMouseDown = false;
@@ -77,17 +68,17 @@ URipple.prototype._down = function (event) {
  * @private
  */
 URipple.prototype._up = function (event) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   var self = this;
   // Don't fire for the artificial "mouseup" generated by a double-click.
   if (event && event.detail !== 2) {
-    (0, _dom.removeClass)(this._rippleElement, 'is-visible');
+    removeClass(this._rippleElement, 'is-visible');
   }
   // Allow a repaint to occur before removing this class, so the animation
   // shows for tap events, which seem to trigger a mouseup too soon after
   // mousedown.
   window.setTimeout(function () {
-    (0, _dom.removeClass)(self._rippleElement, 'is-visible');
+    removeClass(self._rippleElement, 'is-visible');
   }, 0);
 };
 
@@ -96,7 +87,7 @@ URipple.prototype._up = function (event) {
      * @return {number} the frame count.
      */
 URipple.prototype.getFrameCount = function () {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   return this.frameCount_;
 };
 /**
@@ -104,7 +95,7 @@ URipple.prototype.getFrameCount = function () {
      * @param {number} fC the frame count.
      */
 URipple.prototype.setFrameCount = function (fC) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   this.frameCount_ = fC;
 };
 
@@ -113,7 +104,7 @@ URipple.prototype.setFrameCount = function (fC) {
      * @return {Element} the ripple element.
      */
 URipple.prototype.getRippleElement = function () {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   return this._rippleElement;
 };
 
@@ -123,7 +114,7 @@ URipple.prototype.getRippleElement = function () {
  * @param  {number} newY the new Y coordinate
  */
 URipple.prototype.setRippleXY = function (newX, newY) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   this.x_ = newX;
   this.y_ = newY;
 };
@@ -133,7 +124,7 @@ URipple.prototype.setRippleXY = function (newX, newY) {
  * @param  {boolean} start whether or not this is the start frame.
  */
 URipple.prototype.setRippleStyles = function (start) {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   if (this._rippleElement !== null) {
     var transformString;
     var scale;
@@ -155,9 +146,9 @@ URipple.prototype.setRippleStyles = function (start) {
     this._rippleElement.style.transform = transformString;
 
     if (start) {
-      (0, _dom.removeClass)(this._rippleElement, 'is-animating');
+      removeClass(this._rippleElement, 'is-animating');
     } else {
-      (0, _dom.addClass)(this._rippleElement, 'is-animating');
+      addClass(this._rippleElement, 'is-animating');
     }
   }
 };
@@ -166,7 +157,7 @@ URipple.prototype.setRippleStyles = function (start) {
    * Handles an animation frame.
    */
 URipple.prototype.animFrameHandler = function () {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   if (this.frameCount_-- > 0) {
     window.requestAnimationFrame(this.animFrameHandler.bind(this));
   } else {
@@ -178,13 +169,13 @@ URipple.prototype.animFrameHandler = function () {
  * Initialize element.
  */
 URipple.prototype.init = function () {
-  if (_env.env.isIE8 || _env.env.isMobile || _env.env.isAndroidPAD || _env.env.isIPAD) return;
+  if (env.isIE8 || env.isMobile || env.isAndroidPAD || env.isIPAD) return;
   var self = this;
   if (this._element) {
     this._rippleElement = this._element.querySelector('.u-ripple');
     if (!this._rippleElement) {
       this._rippleElement = document.createElement('span');
-      (0, _dom.addClass)(this._rippleElement, 'u-ripple');
+      addClass(this._rippleElement, 'u-ripple');
       this._element.appendChild(this._rippleElement);
       this._element.style.overflow = 'hidden';
       this._element.style.position = 'relative';
@@ -198,23 +189,23 @@ URipple.prototype.init = function () {
     // second ripples. To avoid that, we use this property to ignore the first
     // mouse down after a touch start.
     this._ignoringMouseDown = false;
-    (0, _event.on)(this._element, 'mousedown', function (e) {
+    on(this._element, 'mousedown', function (e) {
       self._down(e);
     });
-    (0, _event.on)(this._element, 'touchstart', function (e) {
+    on(this._element, 'touchstart', function (e) {
       self._down(e);
     });
 
-    (0, _event.on)(this._element, 'mouseup', function (e) {
+    on(this._element, 'mouseup', function (e) {
       self._up(e);
     });
-    (0, _event.on)(this._element, 'mouseleave', function (e) {
+    on(this._element, 'mouseleave', function (e) {
       self._up(e);
     });
-    (0, _event.on)(this._element, 'touchend', function (e) {
+    on(this._element, 'touchend', function (e) {
       self._up(e);
     });
-    (0, _event.on)(this._element, 'blur', function (e) {
+    on(this._element, 'blur', function (e) {
       self._up(e);
     });
   }
@@ -222,5 +213,4 @@ URipple.prototype.init = function () {
 
 var Ripple = URipple;
 
-exports.Ripple = Ripple;
-exports.URipple = URipple;
+export { Ripple, URipple };
