@@ -1,15 +1,10 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
 /**
  * Module : Sparrow class
  * Author : Kvkens(yueming@yonyou.com)
  * Date	  : 2016-07-28 08:45:39
  */
 
-var Class = function Class(o) {
+var Class = function (o) {
 	if (!(this instanceof Class) && isFunction(o)) {
 		return classify(o);
 	}
@@ -94,7 +89,7 @@ function classify(cls) {
 // Mutators define special properties.
 Class.Mutators = {
 
-	'Extends': function Extends(parent) {
+	'Extends': function (parent) {
 		var existed = this.prototype;
 		var proto = createProto(parent.prototype);
 
@@ -112,7 +107,7 @@ Class.Mutators = {
 		this.superclass = parent.prototype;
 	},
 
-	'Implements': function Implements(items) {
+	'Implements': function (items) {
 		isArray(items) || (items = [items]);
 		var proto = this.prototype,
 		    item;
@@ -122,7 +117,7 @@ Class.Mutators = {
 		}
 	},
 
-	'Statics': function Statics(staticProperties) {
+	'Statics': function (staticProperties) {
 		mix(this, staticProperties);
 	}
 };
@@ -163,11 +158,11 @@ var isArray = Array.isArray || function (val) {
 	return toString.call(val) === '[object Array]';
 };
 
-var isFunction = function isFunction(val) {
+var isFunction = function (val) {
 	return toString.call(val) === '[object Function]';
 };
 
-var indexOf = function indexOf(arr, item) {
+var indexOf = function (arr, item) {
 	if (Array.prototype.indexOf && arr.indexOf) {
 		return arr.indexOf(item);
 	} else {
@@ -180,5 +175,4 @@ var indexOf = function indexOf(arr, item) {
 	}
 };
 
-exports.Class = Class;
-exports.isFunction = isFunction;
+export { Class, isFunction };
