@@ -1,15 +1,10 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
 /**
  * Module : Sparrow cookies
  * Author : Kvkens(yueming@yonyou.com)
  * Date	  : 2016-07-27 21:46:50
  */
 
-var setCookie = function setCookie(sName, sValue, oExpires, sPath, sDomain, bSecure) {
+var setCookie = function (sName, sValue, oExpires, sPath, sDomain, bSecure) {
 	var sCookie = sName + "=" + encodeURIComponent(sValue);
 	if (oExpires) sCookie += "; expires=" + oExpires.toGMTString();
 	if (sPath) sCookie += "; path=" + sPath;
@@ -18,7 +13,7 @@ var setCookie = function setCookie(sName, sValue, oExpires, sPath, sDomain, bSec
 	document.cookie = sCookie;
 };
 
-var getCookie = function getCookie(sName) {
+var getCookie = function (sName) {
 	var sRE = "(?:; )?" + sName + "=([^;]*);?";
 	var oRE = new RegExp(sRE);
 
@@ -27,5 +22,4 @@ var getCookie = function getCookie(sName) {
 	} else return null;
 };
 
-exports.setCookie = setCookie;
-exports.getCookie = getCookie;
+export { setCookie, getCookie };
