@@ -1,8 +1,3 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
 /**
  * Module : Sparrow extend enum
  * Author : Kvkens(yueming@yonyou.com)
@@ -20,14 +15,10 @@ var enumerables = true,
 },
     toString = Object.prototype.toString;
 for (var i in enumerablesTest) {
-	exports.enumerables = enumerables = null;
+	enumerables = null;
 }
 if (enumerables) {
-	exports.enumerables = enumerables = ['hasOwnProperty', 'valueOf', 'isPrototypeOf', 'propertyIsEnumerable', 'toLocaleString', 'toString', 'constructor'];
+	enumerables = ['hasOwnProperty', 'valueOf', 'isPrototypeOf', 'propertyIsEnumerable', 'toLocaleString', 'toString', 'constructor'];
 }
 
-exports.enumerables = enumerables;
-exports.U_LANGUAGES = U_LANGUAGES;
-exports.U_THEME = U_THEME;
-exports.U_LOCALE = U_LOCALE;
-exports.U_USERCODE = U_USERCODE;
+export { enumerables, U_LANGUAGES, U_THEME, U_LOCALE, U_USERCODE };
