@@ -1,19 +1,13 @@
-'use strict';
+/**
+ * Module : Sparrow browser environment
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-27 21:46:50
+ */
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.env = undefined;
+import { extend } from './extend';
+var u = {};
 
-var _extend = require('./extend');
-
-var u = {}; /**
-             * Module : Sparrow browser environment
-             * Author : Kvkens(yueming@yonyou.com)
-             * Date	  : 2016-07-27 21:46:50
-             */
-
-(0, _extend.extend)(u, {
+extend(u, {
 	isIE: false,
 	isFF: false,
 	isOpera: false,
@@ -201,4 +195,4 @@ var u = {}; /**
 })();
 
 var env = u;
-exports.env = env;
+export { env };
