@@ -1,36 +1,29 @@
-"use strict";
+/**
+ * Module : Sparrow touch event
+ * Author : Kvkens(yueming@yonyou.com)
+ * Date	  : 2016-07-28 14:41:17
+ */
+import { env } from './env';
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.event = exports.stopEvent = exports.trigger = exports.off = exports.on = undefined;
-
-var _env = require("./env");
-
-var u = {}; /**
-             * Module : Sparrow touch event
-             * Author : Kvkens(yueming@yonyou.com)
-             * Date	  : 2016-07-28 14:41:17
-             */
-
+var u = {};
 u.event = {};
 
-var touchStartEvent = _env.env.hasTouch ? "touchstart" : "mousedown",
-    touchStopEvent = _env.env.hasTouch ? "touchend" : "mouseup",
-    touchMoveEvent = _env.env.hasTouch ? "touchmove" : "mousemove";
+var touchStartEvent = env.hasTouch ? "touchstart" : "mousedown",
+    touchStopEvent = env.hasTouch ? "touchend" : "mouseup",
+    touchMoveEvent = env.hasTouch ? "touchmove" : "mousemove";
 
 // tap和taphold
 u.event.tap = {
 	tapholdThreshold: 750,
 	emitTapOnTaphold: true,
-	touchstartFun: function touchstartFun() {
+	touchstartFun: function () {
 		trigger(this, 'vmousedown');
 	},
-	touchendFun: function touchendFun() {
+	touchendFun: function () {
 		trigger(this, 'vmouseup');
 		trigger(this, 'vclick');
 	},
-	setup: function setup() {
+	setup: function () {
 		var thisObject = this,
 		    isTaphold = false;
 
@@ -82,7 +75,7 @@ u.event.tap = {
 		on(thisObject, 'touchstart', u.event.tap.touchstartFun);
 		on(thisObject, 'touchend', u.event.tap.touchendFun);
 	},
-	teardown: function teardown() {
+	teardown: function () {
 		off(thisObject, 'vmousedown');
 		off(thisObject, 'vclick');
 		off(thisObject, 'vmouseup');
@@ -106,7 +99,7 @@ u.event.swipe = {
 	// Swipe vertical displacement must be less than this.
 	verticalDistanceThreshold: 30,
 
-	getLocation: function getLocation(event) {
+	getLocation: function (event) {
 		var winPageX = window.pageXOffset,
 		    winPageY = window.pageYOffset,
 		    x = event.clientX,
@@ -133,7 +126,7 @@ u.event.swipe = {
 		};
 	},
 
-	start: function start(event) {
+	start: function (event) {
 		var data = event.touches ? event.touches[0] : event,
 		    location = u.event.swipe.getLocation(data);
 		return {
@@ -143,7 +136,7 @@ u.event.swipe = {
 		};
 	},
 
-	stop: function stop(event) {
+	stop: function (event) {
 		var data = event.touches ? event.touches[0] : event,
 		    location = u.event.swipe.getLocation(data);
 		return {
@@ -152,7 +145,7 @@ u.event.swipe = {
 		};
 	},
 
-	handleSwipe: function handleSwipe(start, stop, thisObject, origTarget) {
+	handleSwipe: function (start, stop, thisObject, origTarget) {
 		if (stop.time - start.time < u.event.swipe.durationThreshold && Math.abs(start.coords[0] - stop.coords[0]) > u.event.swipe.horizontalDistanceThreshold && Math.abs(start.coords[1] - stop.coords[1]) < u.event.swipe.verticalDistanceThreshold) {
 			var direction = start.coords[0] > stop.coords[0] ? "swipeleft" : "swiperight";
 
@@ -167,7 +160,7 @@ u.event.swipe = {
 	// in work at any given time
 	eventInProgress: false,
 
-	setup: function setup() {
+	setup: function () {
 		var events,
 		    thisObject = this,
 		    context = {};
@@ -232,7 +225,7 @@ u.event.swipe = {
 		on(thisObject, touchStartEvent, context.start);
 	},
 
-	teardown: function teardown() {
+	teardown: function () {
 		var events, context;
 
 		events = thisObject["mobile-events"];
@@ -265,13 +258,13 @@ u.event.swiperight = u.event.swipe;
 
 var event = u.event;
 
-var on = function on(element, eventName, child, listener) {
+var on = function (element, eventName, child, listener) {
 	if (!element) return;
 	if (arguments.length < 4) {
 		listener = child;
 		child = undefined;
 	} else {
-		var childlistener = function childlistener(e) {
+		var childlistener = function (e) {
 			if (!e) {
 				return;
 			}
@@ -330,7 +323,7 @@ var on = function on(element, eventName, child, listener) {
 	}
 };
 
-var off = function off(element, eventName, listener) {
+var off = function (element, eventName, listener) {
 	//删除事件数组
 	if (listener) {
 		if (element && element["uEvent"] && element["uEvent"][eventName]) {
@@ -361,7 +354,7 @@ var off = function off(element, eventName, listener) {
 	if (element && element["uEvent"] && element["uEvent"][eventName]) element["uEvent"][eventName] = undefined;
 	if (element && element["uEvent"] && element["uEvent"][eventName + 'fn']) element["uEvent"][eventName + 'fn'] = undefined;
 };
-var trigger = function trigger(element, eventName) {
+var trigger = function (element, eventName) {
 	if (element["uEvent"] && element["uEvent"][eventName]) {
 		element["uEvent"][eventName + 'fn']();
 	}
@@ -370,7 +363,7 @@ var trigger = function trigger(element, eventName) {
 /**
  * 阻止冒泡
  */
-var stopEvent = function stopEvent(e) {
+var stopEvent = function (e) {
 	if (typeof e != "undefined") {
 		if (e.stopPropagation) e.stopPropagation();else {
 			e.cancelBubble = true;
@@ -382,8 +375,4 @@ var stopEvent = function stopEvent(e) {
 	}
 };
 
-exports.on = on;
-exports.off = off;
-exports.trigger = trigger;
-exports.stopEvent = stopEvent;
-exports.event = event;
+export { on, off, trigger, stopEvent, event };
