@@ -1,11 +1,3 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
-
 /**
  * Module : Sparrow util tools
  * Author : Kvkens(yueming@yonyou.com)
@@ -16,18 +8,18 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
  * 创建一个带壳的对象,防止外部修改
  * @param {Object} proto
  */
-var createShellObject = function createShellObject(proto) {
-	var exf = function exf() {};
+var createShellObject = function (proto) {
+	var exf = function () {};
 	exf.prototype = proto;
 	return new exf();
 };
-var execIgnoreError = function execIgnoreError(a, b, c) {
+var execIgnoreError = function (a, b, c) {
 	try {
 		a.call(b, c);
 	} catch (e) {}
 };
 
-var getFunction = function getFunction(target, val) {
+var getFunction = function (target, val) {
 	if (!val || typeof val == 'function') return val;
 	if (typeof target[val] == 'function') return target[val];else if (typeof window[val] == 'function') return window[val];else if (val.indexOf('.') != -1) {
 		var func = getJSObject(target, val);
@@ -37,11 +29,11 @@ var getFunction = function getFunction(target, val) {
 	}
 	return val;
 };
-var getJSObject = function getJSObject(target, names) {
+var getJSObject = function (target, names) {
 	if (!names) {
 		return;
 	}
-	if ((typeof names === 'undefined' ? 'undefined' : _typeof(names)) == 'object') return names;
+	if (typeof names == 'object') return names;
 	var nameArr = names.split('.');
 	var obj = target;
 	for (var i = 0; i < nameArr.length; i++) {
@@ -50,10 +42,10 @@ var getJSObject = function getJSObject(target, names) {
 	}
 	return obj;
 };
-var isDate = function isDate(input) {
+var isDate = function (input) {
 	return Object.prototype.toString.call(input) === '[object Date]' || input instanceof Date;
 };
-var isNumber = function isNumber(obj) {
+var isNumber = function (obj) {
 	//return obj === +obj
 	//加了个typeof 判断，因为'431027199110.078573'会解析成number
 	return obj - parseFloat(obj) + 1 >= 0;
@@ -61,14 +53,14 @@ var isNumber = function isNumber(obj) {
 var isArray = Array.isArray || function (val) {
 	return Object.prototype.toString.call(val) === '[object Array]';
 };
-var isEmptyObject = function isEmptyObject(obj) {
+var isEmptyObject = function (obj) {
 	var name;
 	for (name in obj) {
 		return false;
 	}
 	return true;
 };
-var inArray = function inArray(node, arr) {
+var inArray = function (node, arr) {
 	if (!arr instanceof Array) {
 		throw "arguments is not Array";
 	}
@@ -79,14 +71,14 @@ var inArray = function inArray(node, arr) {
 	}
 	return false;
 };
-var isDomElement = function isDomElement(obj) {
+var isDomElement = function (obj) {
 	if (window['HTMLElement']) {
 		return obj instanceof HTMLElement;
 	} else {
 		return obj && obj.tagName && obj.nodeType === 1;
 	}
 };
-var each = function each(obj, callback) {
+var each = function (obj, callback) {
 	if (obj.forEach) {
 		obj.forEach(function (v, k) {
 			callback(k, v);
@@ -121,7 +113,7 @@ String.prototype.replaceAll = function (AFindText, ARepText) {
 	return this.replace(raRegExp, ARepText);
 };
 
-var dateFormat = function dateFormat(str) {
+var dateFormat = function (str) {
 	//如果不是string类型  原型返回
 	if (typeof str !== 'string') {
 		return str;
@@ -143,15 +135,4 @@ var dateFormat = function dateFormat(str) {
 	return str;
 };
 
-exports.createShellObject = createShellObject;
-exports.execIgnoreError = execIgnoreError;
-exports.getFunction = getFunction;
-exports.getJSObject = getJSObject;
-exports.isDate = isDate;
-exports.isNumber = isNumber;
-exports.isArray = isArray;
-exports.isEmptyObject = isEmptyObject;
-exports.inArray = inArray;
-exports.isDomElement = isDomElement;
-exports.each = each;
-exports.dateFormat = dateFormat;
+export { createShellObject, execIgnoreError, getFunction, getJSObject, isDate, isNumber, isArray, isEmptyObject, inArray, isDomElement, each, dateFormat };
