@@ -1,21 +1,13 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.trans = undefined;
-
-var _cookies = require('../cookies');
-
-var _enumerables = require('../enumerables');
-
-// 从datatable/src/compatiable/u/JsExtension.js抽取
 /**
 * Module : Sparrow i18n
 * Author : Kvkens(yueming@yonyou.com)
 * Date	  : 2016-07-29 10:16:54
 */
 //import {uuii18n} from '?';//缺失故修改为default值
+import { getCookie, setCookie } from '../cookies';
+import { U_LOCALE } from '../enumerables';
+
+// 从datatable/src/compatiable/u/JsExtension.js抽取
 window.getCurrentJsPath = function () {
 	var doc = document,
 	    a = {},
@@ -62,13 +54,13 @@ if (window.i18n) {
 		getAsync: false,
 		fallbackLng: false,
 		ns: { namespaces: ['uui-trans'] },
-		lng: (0, _cookies.getCookie)(_enumerables.U_LOCALE) || 'zh',
+		lng: getCookie(U_LOCALE) || 'zh',
 		resGetPath: resGetPath
 	});
 }
 
-var trans = function trans(key, dftValue) {
+var trans = function (key, dftValue) {
 	return window.i18n ? i18n.t('uui-trans:' + key) : dftValue;
 };
 
-exports.trans = trans;
+export { trans };
