@@ -1,8 +1,3 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
 /**
 * Module : Sparrow rsa utils
 * Author : Kvkens(yueming@yonyou.com)
@@ -36,7 +31,7 @@ var maxDigits;
 var ZERO_ARRAY;
 var bigZero, bigOne;
 
-var BigInt = function BigInt(flag) {
+var BigInt = function (flag) {
     if (typeof flag == "boolean" && flag == true) {
         this.digits = null;
     } else {
@@ -48,9 +43,8 @@ var BigInt = function BigInt(flag) {
 RSAUtils.setMaxDigits = function (value) {
     maxDigits = value;
     ZERO_ARRAY = new Array(maxDigits);
-    for (var iza = 0; iza < ZERO_ARRAY.length; iza++) {
-        ZERO_ARRAY[iza] = 0;
-    }bigZero = new BigInt();
+    for (var iza = 0; iza < ZERO_ARRAY.length; iza++) ZERO_ARRAY[iza] = 0;
+    bigZero = new BigInt();
     bigOne = new BigInt();
     bigOne.digits[0] = 1;
 };
@@ -80,9 +74,8 @@ RSAUtils.biFromDecimal = function (s) {
     var i = isNeg ? 1 : 0;
     var result;
     // Skip leading zeros.
-    while (i < s.length && s.charAt(i) == '0') {
-        ++i;
-    }if (i == s.length) {
+    while (i < s.length && s.charAt(i) == '0') ++i;
+    if (i == s.length) {
         result = new BigInt();
     } else {
         var digitCount = s.length - i;
@@ -284,9 +277,8 @@ RSAUtils.biSubtract = function (x, y) {
 
 RSAUtils.biHighIndex = function (x) {
     var result = x.digits.length - 1;
-    while (result > 0 && x.digits[result] == 0) {
-        --result;
-    }return result;
+    while (result > 0 && x.digits[result] == 0) --result;
+    return result;
 };
 
 RSAUtils.biNumBits = function (x) {
@@ -541,7 +533,7 @@ RSAUtils.biPowMod = function (x, y, m) {
     return result;
 };
 
-var BarrettMu = function BarrettMu(m) {
+var BarrettMu = function (m) {
     this.modulus = RSAUtils.biCopy(m);
     this.k = RSAUtils.biHighIndex(this.modulus) + 1;
     var b2k = new BigInt();
@@ -597,7 +589,7 @@ function BarrettMu_powMod(x, y) {
     return result;
 }
 
-var RSAKeyPair = function RSAKeyPair(encryptionExponent, decryptionExponent, modulus) {
+var RSAKeyPair = function (encryptionExponent, decryptionExponent, modulus) {
     var $dmath = RSAUtils;
     this.e = $dmath.biFromHex(encryptionExponent);
     this.d = $dmath.biFromHex(decryptionExponent);
@@ -615,7 +607,7 @@ RSAUtils.getKeyPair = function (encryptionExponent, decryptionExponent, modulus)
     return new RSAKeyPair(encryptionExponent, decryptionExponent, modulus);
 };
 
-var twoDigit = function twoDigit(n) {
+var twoDigit = function (n) {
     return (n < 10 ? "0" : "") + String(n);
 };
 
@@ -695,7 +687,4 @@ RSAUtils.decryptedString = function (options) {
     return text;
 };
 
-exports.RSAUtils = RSAUtils;
-exports.BigInt = BigInt;
-exports.BarrettMu = BarrettMu;
-exports.twoDigit = twoDigit;
+export { RSAUtils, BigInt, BarrettMu, twoDigit };
